requirements.txt
fastapi
uvicorn
langchain
openai
pinecone-client
streamlit
python-dotenv
pypdf
